identity-service/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
cryptography==42.0.5
requests==2.31.0
